Uupdate redis caching example
