+++ docs/source/user/middleware/cache_middleware_async.py
@@ -1,0 +1,17 @@
+import aioredis
+import asyncio
+
+from blacksmith import (
+    AsyncClientFactory,
+    AsyncConsulDiscovery,
+    AsyncHTTPCacheMiddleware,
+)
+
+async def main():
+    cache = aioredis.from_url("redis://redis/0")
+    sd = AsyncConsulDiscovery()
+    cli = AsyncClientFactory(sd).add_middleware(AsyncHTTPCacheMiddleware(cache))
+    await cli.initialize()
+
+
+asyncio.run(main())
