+++ docs/source/user/middleware/cache_middleware_sync.py
@@ -1,7 +1,8 @@
 import redis
 
-from blacksmith import SyncClientFactory, SyncConsulDiscovery, SyncHTTPCachingMiddleware
+from blacksmith import SyncClientFactory, SyncConsulDiscovery, SyncHTTPCacheMiddleware
 
 cache = redis.from_url("redis://redis/0")
 sd = SyncConsulDiscovery()
-cli = SyncClientFactory(sd).add_middleware(SyncHTTPCachingMiddleware(cache))
+cli = SyncClientFactory(sd).add_middleware(SyncHTTPCacheMiddleware(cache))
+cli.initialize()
