+++ examples/redis_caching/notif/pyproject.toml
@@ -7,8 +7,8 @@ authors = ["Guillaume Gauvrit <guillaume@gauvr.it>"]
 [tool.poetry.dependencies]
 python = "^3.9"
 starlette = "^0.17.0"
-uvicorn = "^0.15.0"
-blacksmith = "^0.5.0"
+hypercorn = "^0.13.2"
+blacksmith = { version = "^0.12.1", extras = ["prometheus"] }
 aioredis = {version = "^2.0.0"}
